fix: 修正多处注释

+++ init.sh
@@ -2,23 +2,38 @@
 ###
  # @Author: Ryan
  # @Date: 2021-02-22 20:18:53
- # @LastEditTime: 2021-02-23 15:51:19
+ # @LastEditTime: 2021-02-24 09:01:26
  # @LastEditors: Ryan
- # @Description: VPS初始化脚本
+ # @Description: VPS初始化脚本 For Debian/Ubuntu
  # @FilePath: \VPSReady\init.sh
 ###
 randomNum() {
     awk -v min=10000 -v max=99999 'BEGIN{srand(); print int(min+rand()*(max-min+1))}'
 }
+info () {
+    echo "[I] $*"
+}
+warn () {
+    echo "[W] $*"
+}
+err () {
+    echo "[E] $*"
+}
+suc () {
+    echo "[S] $*"
+}
 if [ -z "$(command -v apt-get)" ]; then
+    err "Only support Debian/Ubuntu"
     exit 1
 fi
 MIRROR=$(echo "${MIRROR-https://raw.githubusercontent.com/benzBrake/VPSReady/main}" | sed 's#/$##g')
 # 0.询问安装内容
 # 1.安装基础软件包
+info "Installing required software"
 apt-get update
 apt-get -y install curl ca-certificates vim
 # 2.设置时区
+info "Modifying timezone"
 if [ -n "$(command -v timedatectl)" ]; then
     timedatectl set-timezone Asia/Shanghai
 else
@@ -26,9 +41,11 @@ else
 fi
 # 3.设置SSH登录参数
 if [ -e "/etc/ssh/sshd_config" ]; then
-    #备份SSH配置
+    # 备份SSH配置
+    info "Backuping SSH config"
     cp -f /etc/ssh/sshd_config /etc/ssh/sshd_config.bak
-    #添加公钥 xiaoji
+    # 添加公钥 xiaoji
+    info "Installing public key"
     mkdir -p /tmp ~/.ssh
     PUBKeyFile="/tmp/$(randomNum).pub"
     while :; do echo
@@ -37,6 +54,7 @@ if [ -e "/etc/ssh/sshd_config" ]; then
     done
     # 有本地用本地（Git clone 项目的时候本地有key），没有就从 Mirror 下载
     if [ -f ./pub/xiaoji.pub ]; then
+        info "Local public key not found, downloading..."
         cp ./pub/xiaoji.pub "${PUBKeyFile}"
     else
         curl -sSL "${MIRROR}/pub/xiaoji.pub" -o "${PUBKeyFile}"
@@ -53,6 +71,7 @@ if [ -e "/etc/ssh/sshd_config" ]; then
         chmod 600 ~/.ssh/authorized_keys
     fi
     # 仅公钥登录
+    info "Enabling only login with public key"
     if grep -i '^PasswordAuthentication\s\+no' /etc/ssh/sshd_config >/dev/null; then
         if grep -i  '^PasswordAuthentication\s\+yes'  /etc/ssh/sshd_config >/dev/null; then
             sed -i "s@^PasswordAuthentication\s\+yes@PasswordAuthentication no@" /etc/ssh/sshd_config
@@ -65,6 +84,7 @@ if [ -e "/etc/ssh/sshd_config" ]; then
         fi
     fi
     # SSH端口
+    info "Changing SSH port to 33022"
     if grep "^[pP][oO][rR][tT]\s\+$aa" /etc/ssh/sshd_config >/dev/null; then
         if grep '#[pP][oO][rR][tT].*' /etc/ssh/sshd_config >/dev/null; then
             sed -i "5aPort 33022" /etc/ssh/sshd_config
@@ -74,23 +94,42 @@ if [ -e "/etc/ssh/sshd_config" ]; then
     else
         sed -i 's@^port.*@Port 33022@i' /etc/ssh/sshd_config
     fi
-    #重启SSH服务
+    # 重启SSH服务
     if [ -n "$(command -v systemctl)" ]; then
-        systemctl restart sshd
+        if systemctl restart sshd; then
+            info "Removing SSH Config Backup..."
+            rm -f /etc/ssh/sshd_config.bak
+        else
+            err "Modify SSH config Failed."
+            info "Restoring SSH Config..."
+            rm -f /etc/ssh/sshd_config
+            mv -f /etc/ssh/sshd_config.bak /etc/ssh/sshd_config
+        fi
     else
         service sshd restart
         service ssh restart
     fi
-    #清理垃圾"
-    rm -rf "${PUBKeyFile}"
+    # 清理临时文件
+    [ -n "${PUBKeyFile}" ] && rm -rf "${PUBKeyFile}"
 else
-    echo "Do not support none OpenSSH Server!"
+    warn "Do not support none OpenSSH Server!"
 fi
-4.安装 Docker
-if [ -f ./.init/docker.sh ]; then
-    chmod +x ./.init/docker.sh
-    ./.init/docker.sh
+# 4.新增用户
+/usr/sbin/useradd -u 1001 mysql
+/usr/sbin/useradd -u 1002 www
+# 5.安装 Docker
+if [ -f /data/init/docker.sh ]; then
+    chmod +x /data/init/docker.sh
+    /data/init/docker.sh
 else
-    bash -c "$(curl -sSL "${MIRROR}/docker.sh" -o -)"
+    bash -c "$(curl -sSL "${MIRROR}/init/docker.sh" -o -)"
 fi
-echo "ALL Done"
+# 创建默认 Docker Compose 配置
+if [ ! -f /data/docker-compose.yml ] && [ -f /data/.docker-compose.yml.demo ]; then
+    if cp -f /data/.docker-compose.yml.demo /data/docker-compose.yml > /dev/null; then
+        info "Create /data/docker-compose.yml"
+    else
+        err "Cannot create /data/docker-compose.yml"
+    fi
+fi
+suc "ALL Done"
